fix: 리포지토리의 ID 타입을 엔티티에 맞게 Long으로 통일

+++ backend/src/main/java/backend/mulkkam/cup/repository/CupRepository.java
@@ -3,5 +3,5 @@ package backend.mulkkam.cup.repository;
 import backend.mulkkam.cup.domain.Cup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CupRepository extends JpaRepository<Cup, Integer> {
+public interface CupRepository extends JpaRepository<Cup, Long> {
 }
