+++ backend/src/main/java/backend/mulkkam/intake/repository/HistoryRepository.java
@@ -3,5 +3,5 @@ package backend.mulkkam.intake.repository;
 import backend.mulkkam.intake.domain.History;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoryRepository extends JpaRepository<History, Integer> {
+public interface HistoryRepository extends JpaRepository<History, Long> {
 }
